moved index to src folder, created a helper function file, implemented queuing system for new connections
